+++ packages/ethereum/sdk/src/order/fill-order/rarible-v2.test.ts
@@ -343,7 +343,23 @@ describe("buy & acceptBid orders", () => {
     )
   })
 
+  test("should match order(buy erc1155 for eth) with dataType=V3", async () => {
+    await testMatchV2OrHigher("RARIBLE_V2_DATA_V3", true)
+  })
+
   test("should match order(buy erc1155 for eth) with dataType=V2", async () => {
+    await testMatchV2OrHigher("RARIBLE_V2_DATA_V2", true)
+  })
+
+  test("should match order(buy erc1155 for eth) with dataType=V3 and no origin fees", async () => {
+    await testMatchV2OrHigher("RARIBLE_V2_DATA_V3", false)
+  })
+
+  test("should match order(buy erc1155 for eth) with dataType=V2 and no origin fees", async () => {
+    await testMatchV2OrHigher("RARIBLE_V2_DATA_V2", false)
+  })
+
+  async function testMatchV2OrHigher(dataType: "RARIBLE_V2_DATA_V2" | "RARIBLE_V2_DATA_V3", includeFees: boolean) {
     await sentTx(it.testErc1155.methods.mint(sellerAddress, 4, 10, "0x"), { from: buyerAddress })
 
     const left: SimpleOrder = {
@@ -365,7 +381,7 @@ describe("buy & acceptBid orders", () => {
       salt: randomWord(),
       type: "RARIBLE_V2",
       data: {
-        dataType: "RARIBLE_V2_DATA_V2",
+        dataType,
         payouts: [],
         originFees: [],
         isMakeFill: true,
@@ -388,7 +404,7 @@ describe("buy & acceptBid orders", () => {
         value: 100,
       },
     ]
-    const tx = await filler.buy({ order: finalOrder, amount: 2, originFees })
+    const tx = await filler.buy({ order: finalOrder, amount: 2, originFees: includeFees ? originFees : [] })
     await tx.wait()
 
     expect(toBn(await it.testErc1155.methods.balanceOf(sellerAddress, 1).call()).toString()).toBe(
@@ -397,7 +413,7 @@ describe("buy & acceptBid orders", () => {
     expect(toBn(await it.testErc1155.methods.balanceOf(buyerAddress, 1).call()).toString()).toBe(
       before1.plus(2).toFixed(),
     )
-  })
+  }
 
   test("should fill order (buy) with crypto punks asset", async () => {
     const punkId = 43